Add compatibility definition for ‘TIMER_IC_POLARITY_BOTH_EDGE’.
This doesn’t exist in some of the versions of the firmware library
supplied by GigaDevice. It’s also not mentioned as a capability of the
chip in the data sheet that I could see. It also uses a reserved bit
of the relevant register.

Assuming I’m right about the above, I would recommend this be
deprecated and removed when possible, or at least marked with some
kind of warning when used as having undefined behavior.