Move assembler startup code into variants directory.
When this code lived in the ‘core/gd32’ directory, all variants were
being automatically compiled and linked in to the resulting ‘core.a’,
which resulted in name collisions and symbols not resolving correctly
for the target architecture unless you got lucky at link time.

Moving the startup code into ‘variants’ resolves this issue, as only
the code matching the variant currently in use is compiled and linked
in.